+++ Sources/Controllers/Map/Widgets/WidgetStates/TripRecordingElevationWidgetState.swift
@@ -85,8 +85,8 @@ final class TripRecordingElevationWidgetState: OAWidgetState {
         }
         
         let pref = OAAppSettings.sharedManager().registerTripRecordingElevationModePreference(prefId, defValue: TripRecordingElevationMode.total.rawValue).makeProfile()
-        if let string = widgetParams?[Self.prefUphillWidgetModeId] as? String, let intVal = Int32(string) {
-            pref.set(Int(intVal))
+        if let string = widgetParams?[Self.prefUphillWidgetModeId] as? String, let intVal = Int(string) {
+            pref.set(intVal)
         }
         
         return pref
