+++ Sources/Controllers/Map/Widgets/WidgetStates/TripRecordingMaxSpeedWidgetState.swift
@@ -80,8 +80,8 @@ final class TripRecordingMaxSpeedWidgetState: OAWidgetState {
         }
         
         let pref = OAAppSettings.sharedManager().registerMaxSpeedModePreference(prefId, defValue: MaxSpeedMode.total.rawValue).makeProfile()
-        if let string = widgetParams?[Self.prefMaxSpeedModeId] as? String, let intVal = Int32(string) {
-            pref.set(Int(intVal))
+        if let string = widgetParams?[Self.prefMaxSpeedModeId] as? String, let intVal = Int(string) {
+            pref.set(intVal)
         }
         
         return pref
